requirements.txt
hydra-core
networkx
numpy
matplotlib
rootutils
# torch packages
torch 
torchvision
# highs MILP solver
highspy
pulp
# hyperparam optimization
optuna